initial commit
This gets things started with a framework for a shared library that
implements the core components of GNOME Builder. Additionally, there is
a command line interface (the foundry program) which allows for interfacing
with those components.

A primary way to interface is through the use of a sub-shell where the
foundry "IDE" is the parent of your shell. For example, to create a new
foundry project inside your project use the init command.

  $ foundry init

This will create a .foundry directory and some skeleton things that we will
use in the future as the tooling is filled out.

To enter a sub-shell use the enter command. It will spawn a sub-shell
hopefully matching the one you are currently in. However, when you run
the foundry command within this sub-shell, your commands are delivered to
the ancestor foundry instance. This persistence allows longer running
workflows to continue behind your shell.

  $ foundry enter
  $ <new sub-shell as child of foundry>

You can tab complete most commands that are available today. Which are not
many since no sub-system is fully ported from Builder.