logging/model: implement log model
This brings over the IdeLogModel with some slight cleanups which map better
to how things are going to get built inside of foundry. The model is still
all in memory, but we may change that in the future to log to disk. For now
that is definitely not an issue given how few messages end up here.