lib/settings: excise the Builder schema path rellics
Since we have a .foundry directory with user and project setting splits,
we do not need to go through the path madness that Builder did to manage
settings within a single namespace.

Instead, we can use separate backends and the same namespaces. This also
gives us one more layer of specificity over what Builder can do today.
Instead this allows 3 layers in the form of:

 * app settings
 * project settings
 * user overrides to project settings

with user overrides being the top priority of the layered settings.