lib/cli: special cause handling of "--"
Everywhere we use that we basically want to say "stop processing here". So
trim that suffix off at the start, and then join it back at the end so we
cannot process it incrementally.

This fixes an issue where "foundry foo -- thing" would lose -- in the argv
due to GOptionContext parsing.