lib/cli: inhibit context shutdown during remote connection
While the remote is connected, inhibit shutdown of the context.

We may need to address this somehow in case we use from Builder and the
build is happening in the command line (and you close the window) as that
would inhibit until completion. But that can be dealt with in other ways.