lib/flatpak: implement Flatpak manifest object in core
Let's be real, we're going to be dealing with these manifests across a
number of tooling bits so it makes sense to have it in the core library
as a file format rather than simply in the plugin.

We may want to create a second pkg-config for this eventually but as long
as we're not linking against libflatpak here we're fine (even though the
plugin may introduce that linkage).