lib/diagnostics: avoid command use for diagnostic tool
This actually ends up easier if we make the diagnostic tool just use an
argv and environ directly.

In doing so, we make the consumer code smaller, and also allow for it to
be overridden when necessary. And to boot, we can check host, pipeline,
and subprocess search paths for the program for robustness.