lib/documentation: add FoundryDocumentationRoot
This describes the base for discoverying documetation and should be
provided by SDK plug-ins. For exmaple, the Flatpak plugin should implement
the DocumentationProvider and yield a listmodel that updates as SDKs are
installed or removed.

A devhelp plug-in should implement the index API to respond to that and
parse/update existing documentation.