plugins/devhelp: import core object skeleton
This is a rewrite of Manuals as a plugin into libfoundry. This doesn't have
the implementation of importers yet but has the scaffolding to do it. That
will come next to bridge into the new documentation API.

The idea is that we could make Drafting, Manuals, and Builder all just
pull in libfoundry and share the same implementation for core components.