lib: remove need for FoundryFutureListModel
This avoids having an intermediate model by instead using data attached
to the GObject GDataList. We do require that only be used from the main
thread (and guards are in place for that).

This has a major benefit in that we can avoid an extra GListModel and
more importantly ensure that GtkSectionModel pass-through when used with
FoundryModelManager.

Ultimately, this makes search in Manuals show the groupings by SDK
properly now.