lib/documentation: tweak query API to preserve section models
This creates a new FoundryDocumentationMatches API which providers may add
their results into. Doing so allows us to both preserve sections naturally
(when using the ModelManager) as well as maintain something simple for just
listing (the DocumentationMatches implements GListModel).

This should hopefully improve the situation with Manuals and wanting to
get those sections propagated to the sidebar properly.