lib/flatpak: ignore flatpak-builder supported fields
In flatpak-builder, it has specific overrides for some properties that will
not result in a parsing failure. Just ignore those as if we suceeded like
it does.

We may need to keep them around if we want to do non-destructive editing
at some point, but for now this is fine.

Fixes: #3