lib/lsp: improve client loading facilities
This ensure the client future is cleared when we fail to init or when
the client exits.

It also ensures that the preferred plugin is used if specified for the
language overrides. If it is not found (such as if it were disabled or
renamed) then a suitable alternative is used.

Of course, if there are no matches, then NULL is returned.