lib/lsp: setup completion request
I'm still somewhat concerned that if we're really going to rely on JSON
for LSPs that we might want to avoid the GVariant conversion and just use
JSON directly. But we can save that for another day and make sure that
we have a compatible variadic API to handle it.