plugins/file-search: implement simple search
This delays fuzzy indexing until the first search is made to avoid extra
startup work. Additionally, it uses a new search result model which
references the index so that we can avoid copying strings or result sets
around too much.

The goal here is to only copy strings when the item is "checked out" of
the list model for display.