plugins/file-search: persist objects after creation
It can cause issues in list model consumption if the pointers at a
particular index change after being requested. So ensure that we keep
the objects alive after creation.

We could use weak pointers here so they are released earlier, but that
is probably a lot more extra work for relatively little real savings.