libfoundry/tests: start on API for unit test extraction/running
The FoundryTest class specifically requires exporting a FoundryCommand so
that we enable ourselves to run them with a tool in the future. For
example, it is often valuable to be able to say "run this test under gdb".

That capability exists in some build systems already (such as Meson), but
we really need our own control so that it can connect to the GDB plugin
within applications rather than just running in a terminal.