libfoundry/git: stub out Git cloning API
I'm going to avoid having an abstracted clone API in foundry and instead
require you use the specific type from the feature. In this case, Git will
export it's own cloner API directly so applications supporting Git
cloning will use that.

Since that is almost certainly the only one people want in applications,
that shouldn't be an issue.