libfoundry/git: implement list_commits_with_file()
This makes it fairly trivial to display a list of commit history for a
file. While the API is convenient, we do have a little annoyance in that
libgit2 does not support the bitmap cache from Git which would accelerate
this operation vastly, similar to [git reflog -- file].

Leave a commit about that in case anyone sees it and wants to work on
that upstream.