libfoundry/platform: replace FoundryAuthPrompt with FoundryInput
This is the first consumer of the new input subsystem.

The input subtypes are generally meant to be thread-safe (though some
effort will be put into making a custom/gtk one in foundry-gtk work in
a safe manner).

To follow this will be making templates use FoundryInput as well.

After that will be work on an IdeTweaks replacement which will also use
this (and add some new types like choices/etc).