libfoundry/llm: add Chat, ChatMessage, and Tool classes
This gives us the API framework to begin a conversation with the LLM model.
No implementation is provided yet and I am curious if we should move the
PluginOllama to a FoundryOllama namespace. Particularly because there will
only be so much we can abstract from the generic model interface. And if
you need more control, you'll want to access the provider-specific APIs.