plugins/gobject-templates: start on gobject code templates
This is something we never got around to in Builder, but the idea of code
templates for one or a small number of files. The first to add is just
a GObject template builder.

I always leaned on the side of snippets for this, but having an actual UI
for doing this could be useful. Additionally, it may be useful to let LLMs
have an API to create the base object and interate rather than start from
scratch on a file (if exposed as a tool).