libfoundry/templates: allow creating directories in templates
This allows you to create an empty directory as part of the template by
specifying a line like this outside of the ``` blocks. It _must_ only
be used after another block has been entered such as ``` or if/end to
help disambiguate between GKeyFile and expansion sections.

The path must end in a / to indicate it is a directory.

But, you can do something like this:

if VersionControl
.git/objects/
.git/refs/heads/
end