plugins/llm-build: add very simple build tool
Ideally we'd have some text file to fill in all of these commands but
lets just start with this for now. Since the subprocess is run within
the project directory, it should connect back to the calling instance
to perform the command. But doing this through the subprocess makes it
convenient to get command output.