libfoundry/llm: add LlmToolCall abstraction
This type represents the single-shot calling of a FoundryLlmTool. The
conversation implementations should create a subclass of this and set it
up to call the tool in the appropriate way based on the communication
style of the model/provider.

One possible use of this is to bind into a conversation history where the
user may authorize a toolcall. We may add additional properties to this
so that descriptions of what will happen can be displayed.

Additionally, the result of the call vfunc should be a LlmMessage. However,
there exists a future where we would want to have messages which can be
more than just text. For example a directory listing could result in a
typed FoundryLlmMessage which also contains a FoundryDirectoryListing. That
way conversations could show more useful information to the user than just
the text which is sent to the model.