libfoundry/tweaks: land most of new tweaks engine
This gets more things in place for managing tweaks in a hierarchical
fashion. This time I kept the FoundryTweakPath internal while also making
it a bit more graceful in what it accespts.

Applications/plugins will just use string paths in the exposed API surface.

To allow for using data tables for the vast majority of internal tweaks,
a TweaksInfo is provided. I've only just started exposing tweaks via this
interface for libfoundry-gtk, but that will continue to take shape soon.

This is expected to build upon FoundryInput so that we don't have to
recreate that multiple times in application/libraries. I expect that at
some point I'll add a libfoundry-adw library which will do that for the
GNOME platform. Other platforms will necessarily need to do that too but
should not be too much effort.