libfoundry-gtk/editing: route more settings through TextSettings
Even though we are only going to change these via the the /app tweak for
the defaults, we can still route them through the text settings so that it
always happens in one place. It also allows applications to override things
if they wish since they can have a TextSettingsProvider which takes
priority over the defaults.