plugins/sarif: add SARIF plugin to receive updates via socket
There is still more that should be done on this, but it gets the basic
receiver in place that we will be able to use from other addins to
redirect diagnostics.

For example, we should be able to set `SARIF_SOCKET` soon with GCC and
direct it here.

I don't love that we're always creating the socket (similar to what we
do with the FoundryDBusService) as they can leak sockets in our
.foundry/tmp/ directory. I'd like to make the socket listener created on
demand once we've requested the address.

This only stores the new diagnostics in a GListStore as part of the
PluginSarifService. We will need an addin to bridge that to the rest
of Foundry so that applications may consume this transparently. For
example, a PluginSarifDiagnosticProvider could filter this model for
diagnostics matching the queried filename.

We might want to add a FoundryDiagnosticProvider.all() vfunc so that
we can get all diagnostics for a "project" regardless of knowing the
filename up-front (such as when the text editor is open).