libfoundry/gir: add FoundryGir parser for .gir
This FoundryGir and FoundryGirNode object structure allows for parsing
*.gir files into a simple in-memory structure.

The goal here is to allow tooling to get better insight into the
documentation found within these files. For example, this could allow
annotating auto-completion with documentation in hover providers that
doesn't require loading an entire WebKitWebView.

Parsing is done on a thread pool along with asynchronous IO to load
the file contents.