libfoundry/files: fix context parsing of grep data
Grep is infinitely tricky, but in this case we have a
useful --null option to help us out to separate the
path from the file data.

Process context as well as we go, dump into batches,
send the batches to the main thread eventually.

It would be nice to cleanup with a thread-safe
implementation of GListStore for batching.