libfoundry/files: abstract file search into grep plugin
This keeps the infrastructure for file search within the libfoundry/files
abstraction but moves the grep implementation into a plugin so that we can
allow for other implementations in the future. That may be helpful in
places where GNU grep is not available (bsdgrep for example which supports
different options) and rg to note two.

This still uses private API from FoundryFileSearchMatch which will be
abstracted in another commit.