libfoundry/terminal: add FoundryTerminalService
While we strictly don't _need_ a service for this, having a service will
allow us to make it very convenient to write applications which allow
spawning a terminal w/ a TerminalIntent via GAction.

This service will collect actions for the common operations and turn those
into intents which will be handled by the application if it intends to
support them.