libfoundry-adw: add FoundryFileRow
This is similar to what was in Builder in terms of use, but very different
in terms of implementation, hopefully for the better.

This just notifies :file whenever there are changes and does the mechanics
of creating the GFile on demand. The file dialog will navigate based on
the user's choice and the GFileType specified for the row.