libfoundry/flatpak: work on non-destructive file-include support
For cases where we are including files from a manifest (such as a sources
or modules list) we need to maintain a non-destructive array of what was
included where. That way we can skip things as we write the JSON document
back out in serialized form.

Granted, this does not allow for us to really do heavy modification
currently to the object structure as we'd lose some of that state in
serialization. But we can approach that problem if/when we make a real
visual manifest editor.

Still to do is the sorting of properties so those remain non-destructive.