libfoundry-adw: remove contextual use for path navigator
We don't want to require the context as a contextual for the navigator
because that implies a GWeakRef and we could start bumping up against the
limits of what GObject supports in terms of max weak-ref count.

In this case, we are fine to just take a full reference to the context.
What I'd like to do next is to remove the need for context on the intent
too and then we can probably drop the context property altogether.