libfoundry/git: add FoundryGitCommitBuilder
The goal of this object is to simplify the process of creating commits
from the current working directory. It tries to provide a much higher
level API that libgit2 directly to simplify implementing in GTK-based
applications.

Commit still needs to be implemented, but this starts to get the
infrastructure in place.