libfoundry/git: provide staged/unstaged delta operations
We really need three things when doing commit building. First we need
access to the untracked files. These will never have deltas since they
are completely new files. The second is the delta from the workdir to
the index (unstaged). Last is the delta from the index to parent commit
(what is staged).

This provides that through the new API

  foundry_git_commit_builder_load_staged_delta()
  foundry_git_commit_builder_load_unstaged_delta()