libfoundry/git: add commit signing support using SSH keys
This adds support for signing commits using ssh-keygen style keys.
The format, while still saying "gpg" has different armor and mildly
binary format within it.

I would like to refactory this shortly to be a method on top of
FoundryGitVcs so that it is easy to reuse in other places where
we want to sign a commit/tag/etc.

Since I sign using SSH keys, this is sort of a must-have. This commit is
signed using this code.